refactor(cli): extract scan dispatcher to scan-command module
F2 reviewer flagged cli.ts at 435 pure LOC after C1, 79 over the 250 cap. Extracts the scan dispatcher (runScan, formatHumanScanSummary, exitCodeForScan, StringWriter) into a dedicated apps/cli/src/scan-command.ts module. cli.ts drops to 271 LOC, just 7 lines over the cap. The scan module owns its own 173-line concern; cli.ts retains the dispatch table and unchanged detect/config show code. 101 tests pass; behavior unchanged.
